contract call proxy separate module check
