+++ contracts/feature-tests/async/forwarder/src/vault_proxy.rs
@@ -1,0 +1,17 @@
+elrond_wasm::imports!();
+
+#[elrond_wasm_derive::callable(VaultProxy)]
+pub trait Vault {
+	fn echo_arguments(
+		&self,
+		args: &VarArgs<BoxedBytes>,
+	) -> ContractCall<BigUint, VarArgs<BoxedBytes>>;
+
+	#[payable("*")]
+	fn accept_funds(&self) -> ContractCall<BigUint, ()>;
+
+	#[payable("*")]
+	fn reject_funds(&self) -> ContractCall<BigUint, ()>;
+
+	fn retrieve_funds(&self, token: TokenIdentifier, amount: BigUint) -> ContractCall<BigUint, ()>;
+}
